150 to go
Most of the issues left are NEI related, and we need to decide how to handle that.
- synced the rider interaction fixes with the 1.7 branch.

+++ src/main/java/ebf/tim/utility/CommonUtil.java
@@ -522,7 +522,7 @@ public class CommonUtil {
             input += 360.0d;
         }
 
-        return input;
+        return (float)input;
     }
 
     /**
